requirements.txt
absl-py==1.0.0
alabaster==0.7.13
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
blis==0.7.9
boto3==1.26.60
botocore==1.29.60
cachetools==5.0.0
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
comm==0.1.2
confection==0.0.4
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distinctipy==1.2.2
docutils==0.19
EditorConfig==0.12.3
ego4d==1.2.5
einops==0.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.6.0
fonttools==4.34.4
fqdn==1.5.1
furo==2022.12.7
future==0.18.2
gdown==4.5.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.44.0
huggingface-hub==0.12.0
idna==3.3
imageio==2.17.0
imagesize==1.4.1
importlib-metadata==4.11.3
ipykernel==6.21.0
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsbeautifier==1.14.4
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langcodes==3.3.0
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.4
-e git+https://github.com/StanfordVL/moma.git@1a91183ff00a7f5ae09ba3a22f4fdf338c640324#egg=momaapi
murmurhash==1.0.9
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8
ninja==1.10.2.3
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pygraphviz==1.10
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytube==12.1.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.8
s3transfer==0.6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sewar==0.4.5
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==6.1.3
sphinx-basic-ng==1.0.0b1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.5
stack-data==0.6.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2022.4.22
timm==0.5.4
tinycss2==1.2.1
tokenizers==0.13.2
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.3.0
uri-template==1.2.0
urllib3==1.26.11
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.1.1
widgetsnbextension==4.0.5
zipp==3.8.0
